+++ tests/test_utils.py
@@ -37,24 +37,30 @@ class TestLoadJson(unittest.TestCase):
 class TestExtractAnswer(unittest.TestCase):
     def test_extract_answer_simple(self):
         app = App.from_json("tests/files/config.json")
+        app.config.inference_prompt = "System\n\n### Input:\n{question}\n\n### Context:\n{context}\n\n### Response:\n"
         app.config.training_prompt = (
             "System\n\n### Input:\n{question}\n\n### Context:\n{context}\n\n### Response:\n{answer}\n\n"
         )
+        tinput_prompt = app.chat_prompt("Question", "Context")
         tinput = app.training_prompt("Question", "Context", "Answer")
 
         expected = "Answer\n\n"
-        result = extract_answer(tinput)
+        result = extract_answer(tinput, tinput_prompt)
         assert result == expected, "Extraction result is not as expected."
 
     def test_extract_answer_simple2(self):
         app = App.from_json("tests/files/config.json")
+        app.config.inference_prompt = (
+            "Goal\n\n### Question:\n{question}\n\n### Additional Info:\n{context}\n\n### Answer:\n"
+        )
         app.config.training_prompt = (
             "Goal\n\n### Question:\n{question}\n\n### Additional Info:\n{context}\n\n### Answer:\n{answer}\n\n"
         )
+        tinput_prompt = app.chat_prompt("Question", "Context")
         tinput = app.training_prompt("Question", "Context", "Answer")
 
-        expected = "Answer\n\n"
-        result = extract_answer(tinput)
+        expected = "Answer"
+        result = extract_answer(tinput, prompt=tinput_prompt, stop="\n\n")
         assert result == expected, "Extraction result is not as expected."